+++ packages/example-app-browser-js/src/index.js
@@ -8,5 +8,5 @@ const app = document.getElementById('app');
 app.innerText = `Hello world from js app! 1 + 1 = ${two} and 2 * 2 = ${four}.`;
 
 if (module.hot) {
-  module.hot.accept(location.reload);
+  module.hot.accept((err) => location.reload());
 }