+++ packages/example-app-browser-ts/src/typings.d.ts
@@ -1,3 +1,1 @@
 declare module 'example-lib-universal-js';
-declare module 'example-lib-universal-ts';
-declare const module;