+++ packages/example-lib-universal-ts/src/index.ts
@@ -2,5 +2,3 @@ import { add } from './add';
 import { subtract } from './subtract';
 
 export { add, subtract };
-
-interface Foo {}